+++ dash_extensions/enrich.py
@@ -14,7 +14,6 @@ import dash
 # noinspection PyUnresolvedReferences
 from dash import (  # lgtm [py/unused-import]
     no_update,
-    Input,
     Output,
     State,
     ClientsideFunction,
@@ -32,7 +31,7 @@ from dash import (  # lgtm [py/unused-import]
     clientside_callback,
 )
 from dash._utils import patch_collections_abc
-from dash.dependencies import _Wildcard
+from dash.dependencies import _Wildcard, DashDependency  # lgtm [py/unused-import]
 from dash.development.base_component import Component
 from flask import session
 from flask_caching.backends import FileSystemCache, RedisCache
@@ -41,8 +40,19 @@ from collections import defaultdict
 from typing import Dict, Callable, List, Union, Any, Tuple
 from datetime import datetime
 
+from dash_extensions import CycleBreaker
+
 _wildcard_mappings = {ALL: "<ALL>", MATCH: "<MATCH>", ALLSMALLER: "<ALLSMALLER>"}
 _wildcard_values = list(_wildcard_mappings.values())
+
+# region Enriched dependencies
+
+class Input(dash.Input):
+    def __init__(self, component_id, component_property, break_cycle=False):
+        super().__init__(component_id, component_property)
+        self.break_cycle = break_cycle
+
+# endregion
 
 # region Dash blueprint
 
@@ -213,6 +223,9 @@ class DashProxy(dash.Dash):
         self.blueprint.register_callbacks(super())
         # Proceed as normally.
         super()._setup_server()
+        # Remap callback bindings to enable callback registration via the 'before_first_request' hook.
+        self.callback = super().callback
+        self.clientside_callback = super().clientside_callback
         # Set session secret. Used by some subclasses.
         if not self.server.secret_key:
             self.server.secret_key = secrets.token_urlsafe(16)
@@ -381,23 +394,24 @@ class BlockingCallbackTransform(DashTransform):
                 "function(){return new Date().getTime();}", Output(end_client_id, "data"), Input(end_server_id, "data")
             )
             # Modify the original callback to send finished signal.
+            single_output = len(callback.outputs) <= 1
             callback.outputs.append(Output(end_server_id, "data"))
             # Modify the original callback to not trigger on inputs, but the new special trigger.
             new_state = [State(item.component_id, item.component_property) for item in callback.inputs]
             callback.inputs = [Input(start_client_id, "data")] + new_state
             # Modify the callback function accordingly.
             f = callback.f
-            callback.f = skip_input_signal_add_output_signal()(f)
+            callback.f = skip_input_signal_add_output_signal(single_output)(f)
 
         return callbacks
 
 
-def skip_input_signal_add_output_signal():
+def skip_input_signal_add_output_signal(single_output: bool):
     def wrapper(f):
         @functools.wraps(f)
         def decorated_function(*args):
             value = f(*args[1:])
-            return _as_list(value) + [datetime.utcnow().timestamp()]
+            return _as_output_list(value, single_output) + [datetime.utcnow().timestamp()]
 
         return decorated_function
 
@@ -509,11 +523,12 @@ class LogTransform(DashTransform):
             if not callback.kwargs.get("log", None):
                 continue
             # Add the log component as output.
+            single_output = len(callback.outputs) <= 1
             callback.outputs.append(self.log_config.log_output)
             # Modify the callback function accordingly.
             f = callback.f
             logger = DashLogger(self.log_config.log_writer_map)  # TODO: What about scope?
-            callback.f = bind_logger(logger)(f)
+            callback.f = bind_logger(logger, single_output)(f)
 
         return callbacks
 
@@ -521,17 +536,57 @@ class LogTransform(DashTransform):
         return [MultiplexerTransform()]
 
 
-def bind_logger(logger):
+def bind_logger(logger, single_output):
     def wrapper(f):
         @functools.wraps(f)
         def decorated_function(*args):
             logger.clear()
             value = f(*args, logger)
-            return _as_list(value) + [logger.get_output()]
+            return _as_output_list(value, single_output) + [logger.get_output()]
 
         return decorated_function
 
     return wrapper
+
+
+# endregion
+
+# region Cycle breaker transform
+
+class CycleBreakerTransform(DashTransform):
+
+    def __init__(self):
+        super().__init__()
+        self.components = []
+
+    def transform_layout(self, layout):
+        children = _as_list(layout.children) + self.components
+        layout.children = children
+
+    def apply(self, callbacks, clientside_callbacks):
+        cycle_inputs = {}
+        # Update inputs.
+        for c in callbacks + clientside_callbacks:
+            for i in c.inputs:
+                if isinstance(i, Input) and i.break_cycle:
+                    cid = self._cycle_break_id(i)
+                    cycle_inputs[cid] = (i.component_id, i.component_property)
+                    i.component_id = cid
+                    i.component_property = "dst"
+        # Construct components.
+        self.components = [CycleBreaker(id=cid) for cid in cycle_inputs]
+        # Construct callbacks.
+        f = "function(x){return x;}"
+        cycle_callbacks = []
+        for cid in cycle_inputs:
+            cb = CallbackBlueprint(Output(cid, "src"), Input(*cycle_inputs[cid]))
+            cb.f = f
+            cycle_callbacks.append(cb)
+        return callbacks, clientside_callbacks + cycle_callbacks
+
+    @staticmethod
+    def _cycle_break_id(d: DashDependency):
+        return f"{str(d).replace('.', '_')}_breaker"
 
 
 # endregion
@@ -1089,6 +1144,7 @@ class Dash(DashProxy):
             LogTransform(),
             MultiplexerTransform(),
             NoOutputTransform(),
+            CycleBreakerTransform(),
             BlockingCallbackTransform(),
             ServersideOutputTransform(**output_defaults),
         ]
@@ -1107,6 +1163,12 @@ def _as_list(item):
     if isinstance(item, list):
         return item
     return [item]
+
+
+def _as_output_list(item, single_output: bool):
+    if single_output:
+        return [item]
+    return _as_list(item)
 
 
 def _create_callback_id(item):